RUST-1800 Modularize unified_runner/operation.rs
