+++ src/test/spec/unified_runner/operation.rs
@@ -2,10 +2,26 @@ mod bulk_write;
 #[cfg(feature = "in-use-encryption")]
 mod csfle;
 mod search_index;
+mod insert_operations;
+mod delete_operations;
+mod find_operations;
+mod failpoint_operations;
+mod update_operations;
+mod list_operations;
+mod collection_operations;
+mod session_operations;
+mod index_operations;
+mod transaction_operations;
+mod upload_download_operations;
+mod count_operations;
+mod command_operations;
+mod connection_operations;
+mod thread_operations;
+mod wait_operations;
+mod topology_operations;
+mod iteration_operations;
 
 use std::{
-    collections::HashMap,
-    convert::TryInto,
     fmt::Debug,
     ops::Deref,
     panic::{catch_unwind, AssertUnwindSafe},
@@ -13,85 +29,58 @@ use std::{
         atomic::{AtomicBool, Ordering},
         Arc,
     },
-    time::Duration,
 };
 
+use collection_operations::{AssertCollectionExists, AssertCollectionNotExists, CreateCollection, DropCollection, RenameCollection};
+use command_operations::{CreateCommandCursor, RunCommand, RunCursorCommand};
+use connection_operations::{AssertNumberConnectionsCheckedOut, Close};
+use count_operations::{Aggregate, AssertEventCount, CountDocuments, Distinct, EstimatedDocumentCount};
+use delete_operations::{DeleteMany, DeleteOne};
+use failpoint_operations::{FailPointCommand, TargetedFailPoint};
+use find_operations::{CreateFindCursor, Find, FindOne, FindOneAndDelete, FindOneAndReplace, FindOneAndUpdate};
 use futures::{
     future::BoxFuture,
-    io::AsyncReadExt,
-    stream::{StreamExt, TryStreamExt},
-    AsyncWriteExt,
     FutureExt,
 };
+use index_operations::{AssertIndexExists, AssertIndexNotExists, CreateIndex, DropIndex, ListIndexNames, ListIndexes};
+use insert_operations::{InsertMany, InsertOne};
+use iteration_operations::{IterateOnce, IterateUntilDocumentOrError};
+use list_operations::{ListCollectionNames, ListCollections, ListDatabaseNames, ListDatabases};
 use serde::{
     de::{DeserializeOwned, Deserializer},
     Deserialize,
 };
+use session_operations::{AssertDifferentLsidOnLastTwoCommands, AssertSameLsidOnLastTwoCommands, AssertSessionDirty, AssertSessionNotDirty, AssertSessionPinned, AssertSessionTransactionState, AssertSessionUnpinned, EndSession};
+use thread_operations::{RunOnThread, WaitForThread};
 use time::OffsetDateTime;
 use tokio::sync::Mutex;
+use topology_operations::{AssertTopologyType, RecordTopologyDescription};
+use transaction_operations::{AbortTransaction, CommitTransaction, StartTransaction, WithTransaction};
+use update_operations::{ReplaceOne, UpdateMany, UpdateOne};
+use upload_download_operations::{Delete, Download, DownloadByName, Upload};
+use wait_operations::{Wait, WaitForEvent, WaitForPrimaryChange};
 
 use super::{
     results_match,
     Entity,
     EntityMap,
     ExpectError,
-    ExpectedEvent,
     TestCursor,
     TestFileEntity,
     TestRunner,
 };
 
 use crate::{
-    action::Action,
-    bson::{doc, to_bson, Bson, Document},
-    client::session::TransactionState,
+    bson::{doc, Bson, Document},
     error::{ErrorKind, Result},
-    options::{
-        AggregateOptions,
-        ChangeStreamOptions,
-        Collation,
-        CountOptions,
-        CreateCollectionOptions,
-        DeleteOptions,
-        DistinctOptions,
-        DropCollectionOptions,
-        DropIndexOptions,
-        EstimatedDocumentCountOptions,
-        FindOneAndDeleteOptions,
-        FindOneAndReplaceOptions,
-        FindOneAndUpdateOptions,
-        FindOneOptions,
-        FindOptions,
-        GridFsDownloadByNameOptions,
-        GridFsUploadOptions,
-        Hint,
-        IndexOptions,
-        InsertManyOptions,
-        InsertOneOptions,
-        ListCollectionsOptions,
-        ListIndexesOptions,
-        ReadConcern,
-        ReplaceOptions,
-        RunCursorCommandOptions,
-        SelectionCriteria,
-        TransactionOptions,
-        UpdateModifications,
-        UpdateOptions,
-    },
-    runtime,
-    serde_util,
-    test::util::fail_point::FailPoint,
-    Collection,
-    Database,
-    IndexModel,
-    ServerType,
-    TopologyType,
+    options::ChangeStreamOptions,
 };
 
 use bulk_write::*;
 #[cfg(feature = "in-use-encryption")]
 use csfle::*;
 use search_index::*;
+
 
 pub(crate) trait TestOperation: Debug + Send + Sync {
     fn execute_test_runner_operation<'a>(
@@ -486,1773 +475,6 @@ impl Deref for Operation {
 
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct DeleteMany {
-    filter: Document,
-    session: Option<String>,
-    #[serde(flatten)]
-    options: DeleteOptions,
-}
-
-impl TestOperation for DeleteMany {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .delete_many(self.filter.clone())
-                    .with_options(self.options.clone())
-            )
-            .await?;
-            let result = to_bson(&result)?;
-            Ok(Some(result.into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct DeleteOne {
-    filter: Document,
-    session: Option<String>,
-    #[serde(flatten)]
-    options: DeleteOptions,
-}
-
-impl TestOperation for DeleteOne {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .delete_one(self.filter.clone())
-                    .with_options(self.options.clone()),
-            )
-            .await?;
-            let result = to_bson(&result)?;
-            Ok(Some(result.into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Default, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct Find {
-    filter: Document,
-    session: Option<String>,
-    // `FindOptions` cannot be embedded directly because serde doesn't support combining `flatten`
-    // and `deny_unknown_fields`, so its fields are replicated here.
-    allow_disk_use: Option<bool>,
-    allow_partial_results: Option<bool>,
-    batch_size: Option<u32>,
-    comment: Option<Bson>,
-    hint: Option<Hint>,
-    limit: Option<i64>,
-    max: Option<Document>,
-    max_scan: Option<u64>,
-    #[serde(
-        default,
-        rename = "maxTimeMS",
-        deserialize_with = "serde_util::deserialize_duration_option_from_u64_millis"
-    )]
-    max_time: Option<Duration>,
-    min: Option<Document>,
-    no_cursor_timeout: Option<bool>,
-    projection: Option<Document>,
-    read_concern: Option<ReadConcern>,
-    return_key: Option<bool>,
-    show_record_id: Option<bool>,
-    skip: Option<u64>,
-    sort: Option<Document>,
-    collation: Option<Collation>,
-    #[serde(rename = "let")]
-    let_vars: Option<Document>,
-}
-
-impl Find {
-    async fn get_cursor<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> Result<TestCursor> {
-        let collection = test_runner.get_collection(id).await;
-
-        // `FindOptions` is constructed without the use of `..Default::default()` to enforce at
-        // compile-time that any new fields added there need to be considered here.
-        let options = FindOptions {
-            allow_disk_use: self.allow_disk_use,
-            allow_partial_results: self.allow_partial_results,
-            batch_size: self.batch_size,
-            comment: self.comment.clone(),
-            hint: self.hint.clone(),
-            limit: self.limit,
-            max: self.max.clone(),
-            max_scan: self.max_scan,
-            max_time: self.max_time,
-            min: self.min.clone(),
-            no_cursor_timeout: self.no_cursor_timeout,
-            projection: self.projection.clone(),
-            read_concern: self.read_concern.clone(),
-            return_key: self.return_key,
-            show_record_id: self.show_record_id,
-            skip: self.skip,
-            sort: self.sort.clone(),
-            collation: self.collation.clone(),
-            cursor_type: None,
-            max_await_time: None,
-            selection_criteria: None,
-            let_vars: self.let_vars.clone(),
-        };
-        let act = collection.find(self.filter.clone()).with_options(options);
-        match &self.session {
-            Some(session_id) => {
-                let cursor = with_mut_session!(test_runner, session_id, |session| async {
-                    act.session(session).await
-                })
-                .await?;
-                Ok(TestCursor::Session {
-                    cursor,
-                    session_id: session_id.clone(),
-                })
-            }
-            None => {
-                let cursor = act.await?;
-                Ok(TestCursor::Normal(Mutex::new(cursor)))
-            }
-        }
-    }
-}
-
-impl TestOperation for Find {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let result = match self.get_cursor(id, test_runner).await? {
-                TestCursor::Session {
-                    mut cursor,
-                    session_id,
-                } => {
-                    with_mut_session!(test_runner, session_id.as_str(), |s| async {
-                        cursor.stream(s).try_collect::<Vec<Document>>().await
-                    })
-                    .await?
-                }
-                TestCursor::Normal(cursor) => {
-                    let cursor = cursor.into_inner();
-                    cursor.try_collect::<Vec<Document>>().await?
-                }
-                TestCursor::ChangeStream(_) => panic!("get_cursor returned a change stream"),
-                TestCursor::Closed => panic!("get_cursor returned a closed cursor"),
-            };
-            Ok(Some(Bson::from(result).into()))
-        }
-        .boxed()
-    }
-
-    fn returns_root_documents(&self) -> bool {
-        true
-    }
-}
-
-#[derive(Debug, Default, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct CreateFindCursor {
-    // `Find` cannot be embedded directly because serde doesn't support combining `flatten`
-    // and `deny_unknown_fields`, so its fields are replicated here.
-    filter: Document,
-    session: Option<String>,
-    allow_disk_use: Option<bool>,
-    allow_partial_results: Option<bool>,
-    batch_size: Option<u32>,
-    comment: Option<Bson>,
-    hint: Option<Hint>,
-    limit: Option<i64>,
-    max: Option<Document>,
-    max_scan: Option<u64>,
-    #[serde(rename = "maxTimeMS")]
-    max_time: Option<Duration>,
-    min: Option<Document>,
-    no_cursor_timeout: Option<bool>,
-    projection: Option<Document>,
-    read_concern: Option<ReadConcern>,
-    return_key: Option<bool>,
-    show_record_id: Option<bool>,
-    skip: Option<u64>,
-    sort: Option<Document>,
-    collation: Option<Collation>,
-    #[serde(rename = "let")]
-    let_vars: Option<Document>,
-}
-
-impl TestOperation for CreateFindCursor {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let find = Find {
-                filter: self.filter.clone(),
-                session: self.session.clone(),
-                allow_disk_use: self.allow_disk_use,
-                allow_partial_results: self.allow_partial_results,
-                batch_size: self.batch_size,
-                comment: self.comment.clone(),
-                hint: self.hint.clone(),
-                limit: self.limit,
-                max: self.max.clone(),
-                max_scan: self.max_scan,
-                max_time: self.max_time,
-                min: self.min.clone(),
-                no_cursor_timeout: self.no_cursor_timeout,
-                projection: self.projection.clone(),
-                read_concern: self.read_concern.clone(),
-                return_key: self.return_key,
-                show_record_id: self.show_record_id,
-                skip: self.skip,
-                sort: self.sort.clone(),
-                collation: self.collation.clone(),
-                let_vars: self.let_vars.clone(),
-            };
-            let cursor = find.get_cursor(id, test_runner).await?;
-            Ok(Some(Entity::Cursor(cursor)))
-        }
-        .boxed()
-    }
-
-    fn returns_root_documents(&self) -> bool {
-        false
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct InsertMany {
-    documents: Vec<Document>,
-    session: Option<String>,
-    #[serde(flatten)]
-    options: InsertManyOptions,
-}
-
-impl TestOperation for InsertMany {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .insert_many(&self.documents)
-                    .with_options(self.options.clone()),
-            )
-            .await?;
-            let ids: HashMap<String, Bson> = result
-                .inserted_ids
-                .into_iter()
-                .map(|(k, v)| (k.to_string(), v))
-                .collect();
-            let ids = to_bson(&ids)?;
-            Ok(Some(Bson::from(doc! { "insertedIds": ids }).into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct InsertOne {
-    document: Document,
-    session: Option<String>,
-    #[serde(flatten)]
-    options: InsertOneOptions,
-}
-
-impl TestOperation for InsertOne {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .insert_one(self.document.clone())
-                    .with_options(self.options.clone()),
-            )
-            .await?;
-            let result = to_bson(&result)?;
-            Ok(Some(result.into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct UpdateMany {
-    filter: Document,
-    update: UpdateModifications,
-    session: Option<String>,
-    #[serde(flatten)]
-    options: UpdateOptions,
-}
-
-impl TestOperation for UpdateMany {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .update_many(self.filter.clone(), self.update.clone())
-                    .with_options(self.options.clone()),
-            )
-            .await?;
-            let result = to_bson(&result)?;
-            Ok(Some(result.into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct UpdateOne {
-    filter: Document,
-    update: UpdateModifications,
-    #[serde(flatten)]
-    options: UpdateOptions,
-    session: Option<String>,
-}
-
-impl TestOperation for UpdateOne {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .update_one(self.filter.clone(), self.update.clone())
-                    .with_options(self.options.clone()),
-            )
-            .await?;
-            let result = to_bson(&result)?;
-            Ok(Some(result.into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-pub(super) struct Aggregate {
-    pipeline: Vec<Document>,
-    session: Option<String>,
-    #[serde(flatten)]
-    options: AggregateOptions,
-}
-
-impl TestOperation for Aggregate {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let result = match &self.session {
-                Some(session_id) => {
-                    enum AggregateEntity {
-                        Collection(Collection<Document>),
-                        Database(Database),
-                        Other(String),
-                    }
-                    let entity = match test_runner.entities.read().await.get(id).unwrap() {
-                        Entity::Collection(c) => AggregateEntity::Collection(c.clone()),
-                        Entity::Database(d) => AggregateEntity::Database(d.clone()),
-                        other => AggregateEntity::Other(format!("{:?}", other)),
-                    };
-                    with_mut_session!(test_runner, session_id, |session| async {
-                        let mut cursor = match entity {
-                            AggregateEntity::Collection(collection) => {
-                                collection
-                                    .aggregate(self.pipeline.clone())
-                                    .with_options(self.options.clone())
-                                    .session(&mut *session)
-                                    .await?
-                            }
-                            AggregateEntity::Database(db) => {
-                                db.aggregate(self.pipeline.clone())
-                                    .with_options(self.options.clone())
-                                    .session(&mut *session)
-                                    .await?
-                            }
-                            AggregateEntity::Other(debug) => {
-                                panic!("Cannot execute aggregate on {}", &debug)
-                            }
-                        };
-                        cursor.stream(session).try_collect::<Vec<Document>>().await
-                    })
-                    .await?
-                }
-                None => {
-                    let entities = test_runner.entities.read().await;
-                    let cursor = match entities.get(id).unwrap() {
-                        Entity::Collection(collection) => {
-                            collection
-                                .aggregate(self.pipeline.clone())
-                                .with_options(self.options.clone())
-                                .await?
-                        }
-                        Entity::Database(db) => {
-                            db.aggregate(self.pipeline.clone())
-                                .with_options(self.options.clone())
-                                .await?
-                        }
-                        other => panic!("Cannot execute aggregate on {:?}", &other),
-                    };
-                    cursor.try_collect::<Vec<Document>>().await?
-                }
-            };
-            Ok(Some(Bson::from(result).into()))
-        }
-        .boxed()
-    }
-
-    fn returns_root_documents(&self) -> bool {
-        true
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct Distinct {
-    field_name: String,
-    filter: Option<Document>,
-    session: Option<String>,
-    #[serde(flatten)]
-    options: DistinctOptions,
-}
-
-impl TestOperation for Distinct {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .distinct(&self.field_name, self.filter.clone().unwrap_or_default())
-                    .with_options(self.options.clone()),
-            )
-            .await?;
-            Ok(Some(Bson::Array(result).into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct CountDocuments {
-    session: Option<String>,
-    filter: Option<Document>,
-    #[serde(flatten)]
-    options: CountOptions,
-}
-
-impl TestOperation for CountDocuments {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .count_documents(self.filter.clone().unwrap_or_default())
-                    .with_options(self.options.clone()),
-            )
-            .await?;
-            Ok(Some(Bson::Int64(result.try_into().unwrap()).into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct EstimatedDocumentCount {
-    #[serde(flatten)]
-    options: EstimatedDocumentCountOptions,
-}
-
-impl TestOperation for EstimatedDocumentCount {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = collection
-                .estimated_document_count()
-                .with_options(self.options.clone())
-                .await?;
-            Ok(Some(Bson::Int64(result.try_into().unwrap()).into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Default)]
-pub(super) struct FindOne {
-    filter: Option<Document>,
-    options: FindOneOptions,
-}
-
-// TODO RUST-1364: remove this impl and derive Deserialize instead
-impl<'de> Deserialize<'de> for FindOne {
-    fn deserialize<D: Deserializer<'de>>(deserializer: D) -> std::result::Result<Self, D::Error> {
-        #[derive(Deserialize)]
-        struct Helper {
-            filter: Option<Document>,
-            #[serde(flatten)]
-            options: FindOneOptions,
-        }
-
-        let helper = Helper::deserialize(deserializer)?;
-
-        Ok(Self {
-            filter: helper.filter,
-            options: helper.options,
-        })
-    }
-}
-
-impl TestOperation for FindOne {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = collection
-                .find_one(self.filter.clone().unwrap_or_default())
-                .with_options(self.options.clone())
-                .await?;
-            match result {
-                Some(result) => Ok(Some(Bson::from(result).into())),
-                None => Ok(Some(Entity::None)),
-            }
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct ListDatabases {
-    session: Option<String>,
-    #[serde(flatten)]
-    options: crate::db::options::ListDatabasesOptions,
-}
-
-impl TestOperation for ListDatabases {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let client = test_runner.get_client(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                client.list_databases().with_options(self.options.clone()),
-            )
-            .await?;
-            Ok(Some(bson::to_bson(&result)?.into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct ListDatabaseNames {
-    #[serde(flatten)]
-    options: crate::db::options::ListDatabasesOptions,
-}
-
-impl TestOperation for ListDatabaseNames {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let client = test_runner.get_client(id).await;
-            let result = client
-                .list_database_names()
-                .with_options(self.options.clone())
-                .await?;
-            let result: Vec<Bson> = result.iter().map(|s| Bson::String(s.to_string())).collect();
-            Ok(Some(Bson::Array(result).into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct ListCollections {
-    session: Option<String>,
-    #[serde(flatten)]
-    options: ListCollectionsOptions,
-}
-
-impl TestOperation for ListCollections {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let db = test_runner.get_database(id).await;
-            let result = match &self.session {
-                Some(session_id) => {
-                    with_mut_session!(test_runner, session_id, |session| async {
-                        let mut cursor = db
-                            .list_collections()
-                            .with_options(self.options.clone())
-                            .session(&mut *session)
-                            .await?;
-                        cursor.stream(session).try_collect::<Vec<_>>().await
-                    })
-                    .await?
-                }
-                None => {
-                    let cursor = db
-                        .list_collections()
-                        .with_options(self.options.clone())
-                        .await?;
-                    cursor.try_collect::<Vec<_>>().await?
-                }
-            };
-            Ok(Some(bson::to_bson(&result)?.into()))
-        }
-        .boxed()
-    }
-
-    fn returns_root_documents(&self) -> bool {
-        true
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct ListCollectionNames {
-    filter: Option<Document>,
-}
-
-impl TestOperation for ListCollectionNames {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let db = test_runner.get_database(id).await;
-            let result = db
-                .list_collection_names()
-                .optional(self.filter.clone(), |b, f| b.filter(f))
-                .await?;
-            let result: Vec<Bson> = result.iter().map(|s| Bson::String(s.to_string())).collect();
-            Ok(Some(Bson::from(result).into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct ReplaceOne {
-    filter: Document,
-    replacement: Document,
-    session: Option<String>,
-    #[serde(flatten)]
-    options: ReplaceOptions,
-}
-
-impl TestOperation for ReplaceOne {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .replace_one(self.filter.clone(), self.replacement.clone())
-                    .with_options(self.options.clone())
-            )
-            .await?;
-            let result = to_bson(&result)?;
-            Ok(Some(result.into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct FindOneAndUpdate {
-    filter: Document,
-    update: UpdateModifications,
-    session: Option<String>,
-    #[serde(flatten)]
-    options: FindOneAndUpdateOptions,
-}
-
-impl TestOperation for FindOneAndUpdate {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .find_one_and_update(self.filter.clone(), self.update.clone())
-                    .with_options(self.options.clone()),
-            )
-            .await?;
-            let result = to_bson(&result)?;
-            Ok(Some(result.into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct FindOneAndReplace {
-    filter: Document,
-    replacement: Document,
-    session: Option<String>,
-    #[serde(flatten)]
-    options: FindOneAndReplaceOptions,
-}
-
-impl TestOperation for FindOneAndReplace {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .find_one_and_replace(self.filter.clone(), self.replacement.clone())
-                    .with_options(self.options.clone())
-            )
-            .await?;
-            let result = to_bson(&result)?;
-
-            Ok(Some(result.into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct FindOneAndDelete {
-    filter: Document,
-    session: Option<String>,
-    #[serde(flatten)]
-    options: FindOneAndDeleteOptions,
-}
-
-impl TestOperation for FindOneAndDelete {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .find_one_and_delete(self.filter.clone())
-                    .with_options(self.options.clone())
-            )
-            .await?;
-            let result = to_bson(&result)?;
-            Ok(Some(result.into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct FailPointCommand {
-    fail_point: FailPoint,
-    client: String,
-}
-
-impl TestOperation for FailPointCommand {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let client = test_runner.get_client(&self.client).await;
-            let guard = client
-                .enable_fail_point(self.fail_point.clone())
-                .await
-                .unwrap();
-            test_runner.fail_point_guards.write().await.push(guard);
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct TargetedFailPoint {
-    fail_point: FailPoint,
-    session: String,
-}
-
-impl TestOperation for TargetedFailPoint {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let selection_criteria =
-                with_mut_session!(test_runner, self.session.as_str(), |session| async {
-                    session
-                        .transaction
-                        .pinned_mongos()
-                        .cloned()
-                        .unwrap_or_else(|| panic!("ClientSession not pinned"))
-                })
-                .await;
-            let guard = test_runner
-                .internal_client
-                .enable_fail_point(
-                    self.fail_point
-                        .clone()
-                        .selection_criteria(selection_criteria),
-                )
-                .await
-                .unwrap();
-            test_runner.fail_point_guards.write().await.push(guard);
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertCollectionExists {
-    collection_name: String,
-    database_name: String,
-}
-
-impl TestOperation for AssertCollectionExists {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let db = test_runner.internal_client.database(&self.database_name);
-            let names = db.list_collection_names().await.unwrap();
-            assert!(names.contains(&self.collection_name));
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertCollectionNotExists {
-    collection_name: String,
-    database_name: String,
-}
-
-impl TestOperation for AssertCollectionNotExists {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let db = test_runner.internal_client.database(&self.database_name);
-            let names = db.list_collection_names().await.unwrap();
-            assert!(!names.contains(&self.collection_name));
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct CreateCollection {
-    collection: String,
-    #[serde(flatten)]
-    options: CreateCollectionOptions,
-    session: Option<String>,
-}
-
-impl TestOperation for CreateCollection {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let database = test_runner.get_database(id).await;
-            with_opt_session!(
-                test_runner,
-                &self.session,
-                database
-                    .create_collection(&self.collection)
-                    .with_options(self.options.clone()),
-            )
-            .await?;
-            Ok(Some(Entity::Collection(
-                database.collection(&self.collection),
-            )))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct DropCollection {
-    collection: String,
-    #[serde(flatten)]
-    options: DropCollectionOptions,
-    session: Option<String>,
-}
-
-impl TestOperation for DropCollection {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let database = test_runner.get_database(id).await;
-            let collection = database.collection::<Document>(&self.collection).clone();
-            with_opt_session!(
-                test_runner,
-                &self.session,
-                collection.drop().with_options(self.options.clone()),
-            )
-            .await?;
-            Ok(None)
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct RunCommand {
-    command: Document,
-    // We don't need to use this field, but it needs to be included during deserialization so that
-    // we can use the deny_unknown_fields tag.
-    #[serde(rename = "commandName")]
-    _command_name: String,
-    read_preference: Option<SelectionCriteria>,
-    session: Option<String>,
-}
-
-impl TestOperation for RunCommand {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let command = self.command.clone();
-
-            let db = test_runner.get_database(id).await;
-            let result = with_opt_session!(
-                test_runner,
-                &self.session,
-                db.run_command(command)
-                    .optional(self.read_preference.clone(), |a, rp| {
-                        a.selection_criteria(rp)
-                    }),
-            )
-            .await?;
-            let result = to_bson(&result)?;
-            Ok(Some(result.into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct RunCursorCommand {
-    command: Document,
-    // We don't need to use this field, but it needs to be included during deserialization so that
-    // we can use the deny_unknown_fields tag.
-    #[serde(rename = "commandName")]
-    _command_name: String,
-
-    #[serde(flatten)]
-    options: RunCursorCommandOptions,
-    session: Option<String>,
-}
-
-impl TestOperation for RunCursorCommand {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let command = self.command.clone();
-            let db = test_runner.get_database(id).await;
-            let options = self.options.clone();
-
-            let action = db.run_cursor_command(command).with_options(options);
-            let result = match &self.session {
-                Some(session_id) => {
-                    with_mut_session!(test_runner, session_id, |session| async {
-                        let mut cursor = action.session(&mut *session).await?;
-                        cursor.stream(session).try_collect::<Vec<_>>().await
-                    })
-                    .await?
-                }
-                None => {
-                    let cursor = action.await?;
-                    cursor.try_collect::<Vec<_>>().await?
-                }
-            };
-
-            Ok(Some(bson::to_bson(&result)?.into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub struct CreateCommandCursor {
-    command: Document,
-    // We don't need to use this field, but it needs to be included during deserialization so that
-    // we can use the deny_unknown_fields tag.
-    #[serde(rename = "commandName")]
-    _command_name: String,
-
-    #[serde(flatten)]
-    options: RunCursorCommandOptions,
-    session: Option<String>,
-}
-
-impl TestOperation for CreateCommandCursor {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let command = self.command.clone();
-            let db = test_runner.get_database(id).await;
-            let options = self.options.clone();
-
-            let action = db.run_cursor_command(command).with_options(options);
-            match &self.session {
-                Some(session_id) => {
-                    let mut ses_cursor = None;
-                    with_mut_session!(test_runner, session_id, |session| async {
-                        ses_cursor = Some(action.session(session).await);
-                    })
-                    .await;
-                    let test_cursor = TestCursor::Session {
-                        cursor: ses_cursor.unwrap().unwrap(),
-                        session_id: session_id.clone(),
-                    };
-                    Ok(Some(Entity::Cursor(test_cursor)))
-                }
-                None => {
-                    let doc_cursor = action.await?;
-                    let test_cursor = TestCursor::Normal(Mutex::new(doc_cursor));
-                    Ok(Some(Entity::Cursor(test_cursor)))
-                }
-            }
-        }
-        .boxed()
-    }
-
-    fn returns_root_documents(&self) -> bool {
-        false
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct EndSession {}
-
-impl TestOperation for EndSession {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            match test_runner.entities.write().await.get_mut(id) {
-                Some(Entity::Session(session)) => session.client_session.take(),
-                e => panic!("expected session for {:?}, got {:?}", id, e),
-            };
-            tokio::time::sleep(Duration::from_secs(1)).await;
-            Ok(None)
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertSessionTransactionState {
-    session: String,
-    state: String,
-}
-
-impl TestOperation for AssertSessionTransactionState {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let session_state =
-                with_mut_session!(test_runner, self.session.as_str(), |session| async {
-                    match &session.transaction.state {
-                        TransactionState::None => "none",
-                        TransactionState::Starting => "starting",
-                        TransactionState::InProgress => "in_progress",
-                        TransactionState::Committed { data_committed: _ } => "committed",
-                        TransactionState::Aborted => "aborted",
-                    }
-                })
-                .await;
-            assert_eq!(session_state, self.state);
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertSessionPinned {
-    session: String,
-}
-
-impl TestOperation for AssertSessionPinned {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let is_pinned =
-                with_mut_session!(test_runner, self.session.as_str(), |session| async {
-                    session.transaction.pinned_mongos().is_some()
-                })
-                .await;
-            assert!(is_pinned);
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertSessionUnpinned {
-    session: String,
-}
-
-impl TestOperation for AssertSessionUnpinned {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let is_pinned = with_mut_session!(test_runner, self.session.as_str(), |session| {
-                async move { session.transaction.pinned_mongos().is_some() }
-            })
-            .await;
-            assert!(!is_pinned);
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertDifferentLsidOnLastTwoCommands {
-    client: String,
-}
-
-impl TestOperation for AssertDifferentLsidOnLastTwoCommands {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let entities = test_runner.entities.read().await;
-            let client = entities.get(&self.client).unwrap().as_client();
-            let events = client.get_all_command_started_events();
-
-            let lsid1 = events[events.len() - 1].command.get("lsid").unwrap();
-            let lsid2 = events[events.len() - 2].command.get("lsid").unwrap();
-            assert_ne!(lsid1, lsid2);
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertSameLsidOnLastTwoCommands {
-    client: String,
-}
-
-impl TestOperation for AssertSameLsidOnLastTwoCommands {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let entities = test_runner.entities.read().await;
-            let client = entities.get(&self.client).unwrap().as_client();
-            client.sync_workers().await;
-            let events = client.get_all_command_started_events();
-
-            let lsid1 = events[events.len() - 1].command.get("lsid").unwrap();
-            let lsid2 = events[events.len() - 2].command.get("lsid").unwrap();
-            assert_eq!(lsid1, lsid2);
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertSessionDirty {
-    session: String,
-}
-
-impl TestOperation for AssertSessionDirty {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let dirty = with_mut_session!(test_runner, self.session.as_str(), |session| {
-                async move { session.is_dirty() }.boxed()
-            })
-            .await;
-            assert!(dirty);
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertSessionNotDirty {
-    session: String,
-}
-
-impl TestOperation for AssertSessionNotDirty {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let dirty = with_mut_session!(test_runner, self.session.as_str(), |session| {
-                async move { session.is_dirty() }
-            })
-            .await;
-            assert!(!dirty);
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct StartTransaction {
-    #[serde(flatten)]
-    options: TransactionOptions,
-}
-
-impl TestOperation for StartTransaction {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            with_mut_session!(test_runner, id, |session| {
-                async move {
-                    session
-                        .start_transaction()
-                        .with_options(self.options.clone())
-                        .await
-                }
-            })
-            .await?;
-            Ok(None)
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct CommitTransaction {}
-
-impl TestOperation for CommitTransaction {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            with_mut_session!(test_runner, id, |session| {
-                async move { session.commit_transaction().await }
-            })
-            .await?;
-            Ok(None)
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AbortTransaction {}
-
-impl TestOperation for AbortTransaction {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            with_mut_session!(test_runner, id, |session| async move {
-                session.abort_transaction().await
-            })
-            .await?;
-            Ok(None)
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct WithTransaction {
-    callback: Vec<Operation>,
-    #[serde(flatten)]
-    options: Option<TransactionOptions>,
-}
-
-impl TestOperation for WithTransaction {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            with_mut_session!(test_runner, id, |session| async move {
-                session
-                    .start_transaction()
-                    .with_options(self.options.clone())
-                    .and_run(
-                        (&self.callback, test_runner),
-                        |session, (callback, test_runner)| {
-                            async move {
-                                test_runner.entities.write().await.insert(
-                                    id.to_string(),
-                                    Entity::SessionPtr(super::entity::SessionPtr(session)),
-                                );
-                                let mut result = Ok(());
-                                for op in callback.iter() {
-                                    let r =
-                                        op.execute_fallible(test_runner, "withTransaction").await;
-                                    if r.is_err() {
-                                        result = r;
-                                        break;
-                                    }
-                                }
-                                test_runner.entities.write().await.remove(id);
-                                result
-                            }
-                            .boxed()
-                        },
-                    )
-                    .await
-            })
-            .await?;
-            Ok(None)
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(deny_unknown_fields)]
-pub(super) struct CreateIndex {
-    session: Option<String>,
-    keys: Document,
-    name: Option<String>,
-}
-
-impl TestOperation for CreateIndex {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let options = IndexOptions::builder().name(self.name.clone()).build();
-            let index = IndexModel::builder()
-                .keys(self.keys.clone())
-                .options(options)
-                .build();
-
-            let collection = test_runner.get_collection(id).await;
-            let name =
-                with_opt_session!(test_runner, &self.session, collection.create_index(index))
-                    .await?
-                    .index_name;
-            Ok(Some(Bson::String(name).into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-pub(super) struct ListIndexes {
-    session: Option<String>,
-    #[serde(flatten)]
-    options: ListIndexesOptions,
-}
-
-impl TestOperation for ListIndexes {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let act = collection.list_indexes().with_options(self.options.clone());
-            let indexes: Vec<IndexModel> = match self.session {
-                Some(ref session) => {
-                    with_mut_session!(test_runner, session, |session| {
-                        async {
-                            act.session(&mut *session)
-                                .await?
-                                .stream(session)
-                                .try_collect()
-                                .await
-                        }
-                    })
-                    .await?
-                }
-                None => act.await?.try_collect().await?,
-            };
-            let indexes: Vec<Document> = indexes
-                .iter()
-                .map(|index| bson::to_document(index).unwrap())
-                .collect();
-            Ok(Some(Bson::from(indexes).into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-pub(super) struct ListIndexNames {
-    session: Option<String>,
-}
-
-impl TestOperation for ListIndexNames {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            let names =
-                with_opt_session!(test_runner, &self.session, collection.list_index_names(),)
-                    .await?;
-            Ok(Some(Bson::from(names).into()))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertIndexExists {
-    collection_name: String,
-    database_name: String,
-    index_name: String,
-}
-
-impl TestOperation for AssertIndexExists {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let coll = test_runner
-                .internal_client
-                .database(&self.database_name)
-                .collection::<Document>(&self.collection_name);
-            let names = coll.list_index_names().await.unwrap();
-            assert!(names.contains(&self.index_name));
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertIndexNotExists {
-    collection_name: String,
-    database_name: String,
-    index_name: String,
-}
-
-impl TestOperation for AssertIndexNotExists {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let coll = test_runner
-                .internal_client
-                .database(&self.database_name)
-                .collection::<Document>(&self.collection_name);
-            match coll.list_index_names().await {
-                Ok(indexes) => assert!(!indexes.contains(&self.index_name)),
-                // a namespace not found error indicates that the index does not exist
-                Err(err) => assert_eq!(err.sdam_code(), Some(26)),
-            }
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct IterateUntilDocumentOrError {}
-
-impl TestOperation for IterateUntilDocumentOrError {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            // A `SessionCursor` also requires a `&mut Session`, which would cause conflicting
-            // borrows, so take the cursor from the map and return it after execution instead.
-            let mut cursor = test_runner.take_cursor(id).await;
-            let next = match &mut cursor {
-                TestCursor::Normal(cursor) => {
-                    let mut cursor = cursor.lock().await;
-                    cursor.next().await
-                }
-                TestCursor::Session { cursor, session_id } => {
-                    cursor
-                        .next(
-                            test_runner
-                                .entities
-                                .write()
-                                .await
-                                .get_mut(session_id)
-                                .unwrap()
-                                .as_mut_session(),
-                        )
-                        .await
-                }
-                TestCursor::ChangeStream(stream) => {
-                    let mut stream = stream.lock().await;
-                    stream.next().await.map(|res| {
-                        res.map(|ev| match bson::to_bson(&ev) {
-                            Ok(Bson::Document(doc)) => doc,
-                            _ => panic!("invalid serialization result"),
-                        })
-                    })
-                }
-                TestCursor::Closed => None,
-            };
-            test_runner.return_cursor(id, cursor).await;
-            next.transpose()
-                .map(|opt| opt.map(|doc| Entity::Bson(Bson::Document(doc))))
-        }
-        .boxed()
-    }
-
-    fn returns_root_documents(&self) -> bool {
-        true
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct Close {}
-
-impl TestOperation for Close {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let mut entities = test_runner.entities.write().await;
-            let target_entity = entities.get(id).unwrap();
-            match target_entity {
-                Entity::Client(_) => {
-                    let client = entities.get_mut(id).unwrap().as_mut_client();
-                    let closed_client_topology_id = client.topology_id;
-                    client
-                        .client
-                        .take()
-                        .unwrap()
-                        .shutdown()
-                        .immediate(true)
-                        .await;
-
-                    let mut entities_to_remove = vec![];
-                    for (key, value) in entities.iter() {
-                        match value {
-                            // skip clients so that we don't remove the client entity itself from
-                            // the map: we want to preserve it so we can
-                            // access the other data stored on the entity.
-                            Entity::Client(_) => {}
-                            _ => {
-                                if value.client_topology_id().await
-                                    == Some(closed_client_topology_id)
-                                {
-                                    entities_to_remove.push(key.clone());
-                                }
-                            }
-                        }
-                    }
-                    for entity_id in entities_to_remove {
-                        entities.remove(&entity_id);
-                    }
-
-                    Ok(None)
-                }
-                Entity::Cursor(_) => {
-                    let cursor = entities.get_mut(id).unwrap().as_mut_cursor();
-                    let rx = cursor.make_kill_watcher().await;
-                    *cursor = TestCursor::Closed;
-                    drop(entities);
-                    let _ = rx.await;
-                    Ok(None)
-                }
-                _ => panic!(
-                    "Unsupported entity {:?} for close operation; expected Client or Cursor",
-                    target_entity
-                ),
-            }
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertNumberConnectionsCheckedOut {
-    client: String,
-    connections: u32,
-}
-
-impl TestOperation for AssertNumberConnectionsCheckedOut {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let client = test_runner.get_client(&self.client).await;
-            client.sync_workers().await;
-            assert_eq!(client.connections_checked_out(), self.connections);
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
 pub(super) struct CreateChangeStream {
     pipeline: Vec<Document>,
     #[serde(flatten)]
@@ -2292,35 +514,6 @@ impl TestOperation for CreateChangeStream {
             Ok(Some(Entity::Cursor(TestCursor::ChangeStream(Mutex::new(
                 stream.with_type::<Document>(),
             )))))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct RenameCollection {
-    to: String,
-}
-
-impl TestOperation for RenameCollection {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let target = test_runner.get_collection(id).await;
-            let ns = target.namespace();
-            let mut to_ns = ns.clone();
-            to_ns.coll.clone_from(&self.to);
-            let cmd = doc! {
-                "renameCollection": crate::bson::to_bson(&ns)?,
-                "to": crate::bson::to_bson(&to_ns)?,
-            };
-            let admin = test_runner.internal_client.database("admin");
-            admin.run_command(cmd).await?;
-            Ok(None)
         }
         .boxed()
     }
@@ -2537,201 +730,6 @@ impl Loop {
         }
     }
 }
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct RunOnThread {
-    thread: String,
-    operation: Arc<Operation>,
-}
-
-impl TestOperation for RunOnThread {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async {
-            let thread = test_runner.get_thread(self.thread.as_str()).await;
-            thread.run_operation(self.operation.clone());
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct WaitForThread {
-    thread: String,
-}
-
-impl TestOperation for WaitForThread {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async {
-            let thread = test_runner.get_thread(self.thread.as_str()).await;
-            thread.wait().await.unwrap_or_else(|e| {
-                panic!("thread {:?} did not exit successfully: {}", self.thread, e)
-            });
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertEventCount {
-    client: String,
-    event: ExpectedEvent,
-    count: usize,
-}
-
-impl TestOperation for AssertEventCount {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async {
-            let client = test_runner.get_client(self.client.as_str()).await;
-            let entities = test_runner.entities.read().await;
-            let actual_events = client.matching_events(&self.event, &entities);
-            assert_eq!(
-                actual_events.len(),
-                self.count,
-                "expected to see {} events matching: {:#?}, instead saw: {:#?}",
-                self.count,
-                self.event,
-                actual_events
-            );
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct WaitForEvent {
-    client: String,
-    event: ExpectedEvent,
-    count: usize,
-}
-
-impl TestOperation for WaitForEvent {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async {
-            let client = test_runner.get_client(self.client.as_str()).await;
-            let entities = test_runner.entities.clone();
-            client
-                .wait_for_matching_events(&self.event, self.count, entities)
-                .await
-                .unwrap();
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct RecordTopologyDescription {
-    id: String,
-    client: String,
-}
-
-impl TestOperation for RecordTopologyDescription {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async {
-            let client = test_runner.get_client(&self.client).await;
-            let description = client.topology_description();
-            test_runner.insert_entity(&self.id, description).await;
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct AssertTopologyType {
-    topology_description: String,
-    topology_type: TopologyType,
-}
-
-impl TestOperation for AssertTopologyType {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async {
-            let td = test_runner
-                .get_topology_description(&self.topology_description)
-                .await;
-            assert_eq!(td.topology_type, self.topology_type);
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct WaitForPrimaryChange {
-    client: String,
-    prior_topology_description: String,
-    #[serde(rename = "timeoutMS")]
-    timeout_ms: Option<u64>,
-}
-
-impl TestOperation for WaitForPrimaryChange {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        async move {
-            let client = test_runner.get_client(&self.client).await;
-            let td = test_runner
-                .get_topology_description(&self.prior_topology_description)
-                .await;
-            let old_primary = td.servers_with_type(&[ServerType::RsPrimary]).next();
-            let timeout = Duration::from_millis(self.timeout_ms.unwrap_or(10_000));
-
-            runtime::timeout(timeout, async {
-                let mut watcher = client.topology().watch();
-
-                loop {
-                    let latest = watcher.observe_latest();
-                    if let Some(primary) = latest.description.primary() {
-                        if Some(primary) != old_primary {
-                            return;
-                        }
-                    }
-                    watcher.wait_for_update(None).await;
-                }
-            })
-            .await
-            .unwrap();
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct Wait {
-    ms: u64,
-}
-
-impl TestOperation for Wait {
-    fn execute_test_runner_operation<'a>(
-        &'a self,
-        _test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, ()> {
-        tokio::time::sleep(Duration::from_millis(self.ms)).boxed()
-    }
-}
 
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase", deny_unknown_fields)]
@@ -2752,202 +750,6 @@ impl TestOperation for CreateEntities {
     #[cfg(feature = "tracing-unstable")]
     fn test_file_entities(&self) -> Option<&Vec<TestFileEntity>> {
         Some(&self.entities)
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct Download {
-    id: Bson,
-}
-
-impl TestOperation for Download {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let bucket = test_runner.get_bucket(id).await;
-
-            let mut buf: Vec<u8> = vec![];
-            bucket
-                .open_download_stream(self.id.clone())
-                .await?
-                .read_to_end(&mut buf)
-                .await?;
-            let writer_data = hex::encode(buf);
-
-            Ok(Some(Entity::Bson(writer_data.into())))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct DownloadByName {
-    filename: String,
-    #[serde(flatten)]
-    options: GridFsDownloadByNameOptions,
-}
-
-impl TestOperation for DownloadByName {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let bucket = test_runner.get_bucket(id).await;
-
-            let mut buf: Vec<u8> = vec![];
-            bucket
-                .open_download_stream_by_name(&self.filename)
-                .with_options(self.options.clone())
-                .await?
-                .read_to_end(&mut buf)
-                .await?;
-            let writer_data = hex::encode(buf);
-
-            Ok(Some(Entity::Bson(writer_data.into())))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct Delete {
-    id: Bson,
-}
-
-impl TestOperation for Delete {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let bucket = test_runner.get_bucket(id).await;
-            bucket.delete(self.id.clone()).await?;
-            Ok(None)
-        }
-        .boxed()
-    }
-}
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct Upload {
-    source: Document,
-    filename: String,
-    // content_type and disableMD5 are deprecated and no longer supported.
-    // Options included for deserialization.
-    #[serde(rename = "contentType")]
-    _content_type: Option<String>,
-    #[serde(rename = "disableMD5")]
-    _disable_md5: Option<bool>,
-    #[serde(flatten)]
-    options: GridFsUploadOptions,
-}
-
-impl TestOperation for Upload {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let bucket = test_runner.get_bucket(id).await;
-            let hex_string = self.source.get("$$hexBytes").unwrap().as_str().unwrap();
-            let bytes = hex::decode(hex_string).unwrap();
-
-            let id = {
-                let mut stream = bucket
-                    .open_upload_stream(&self.filename)
-                    .with_options(self.options.clone())
-                    .await?;
-                stream.write_all(&bytes[..]).await?;
-                stream.close().await?;
-                stream.id().clone()
-            };
-
-            Ok(Some(Entity::Bson(id)))
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-#[serde(rename_all = "camelCase", deny_unknown_fields)]
-pub(super) struct IterateOnce {}
-
-impl TestOperation for IterateOnce {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let mut cursor = test_runner.take_cursor(id).await;
-            match &mut cursor {
-                TestCursor::Normal(cursor) => {
-                    let mut cursor = cursor.lock().await;
-                    cursor.try_advance().await?;
-                }
-                TestCursor::Session { cursor, session_id } => {
-                    cursor
-                        .try_advance(
-                            test_runner
-                                .entities
-                                .write()
-                                .await
-                                .get_mut(session_id)
-                                .unwrap()
-                                .as_mut_session(),
-                        )
-                        .await?;
-                }
-                TestCursor::ChangeStream(change_stream) => {
-                    let mut change_stream = change_stream.lock().await;
-                    change_stream.next_if_any().await?;
-                }
-                TestCursor::Closed => panic!("Attempted to call IterateOnce on a closed cursor"),
-            }
-            test_runner.return_cursor(id, cursor).await;
-            Ok(None)
-        }
-        .boxed()
-    }
-}
-
-#[derive(Debug, Deserialize)]
-pub(super) struct DropIndex {
-    session: Option<String>,
-    name: String,
-    #[serde(flatten)]
-    options: Option<DropIndexOptions>,
-}
-
-impl TestOperation for DropIndex {
-    fn execute_entity_operation<'a>(
-        &'a self,
-        id: &'a str,
-        test_runner: &'a TestRunner,
-    ) -> BoxFuture<'a, Result<Option<Entity>>> {
-        async move {
-            let collection = test_runner.get_collection(id).await;
-            with_opt_session!(
-                test_runner,
-                &self.session,
-                collection
-                    .drop_index(&self.name)
-                    .with_options(self.options.clone())
-            )
-            .await?;
-            Ok(None)
-        }
-        .boxed()
     }
 }
