+++ src/test/spec/unified_runner/operation/delete_operations.rs
@@ -1,0 +1,74 @@
+use bson::{to_bson, Document};
+use serde::Deserialize;
+use crate::test::spec::unified_runner::operation::TestOperation;
+use crate::options::DeleteOptions;
+use crate::test::spec::unified_runner::{Entity, TestRunner};
+use futures::future::BoxFuture;
+use crate::test::spec::unified_runner::operation::{with_opt_session, with_mut_session};
+use crate::error::Result;
+use futures_util::FutureExt;
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase", deny_unknown_fields)]
+pub(super) struct DeleteMany {
+    filter: Document,
+    session: Option<String>,
+    #[serde(flatten)]
+    options: DeleteOptions,
+}
+
+impl TestOperation for DeleteMany {
+    fn execute_entity_operation<'a>(
+        &'a self,
+        id: &'a str,
+        test_runner: &'a TestRunner,
+    ) -> BoxFuture<'a, Result<Option<Entity>>> {
+        async move {
+            let collection = test_runner.get_collection(id).await;
+            let result = with_opt_session!(
+                test_runner,
+                &self.session,
+                collection
+                    .delete_many(self.filter.clone())
+                    .with_options(self.options.clone())
+            )
+            .await?;
+            let result = to_bson(&result)?;
+            Ok(Some(result.into()))
+        }
+        .boxed()
+    }
+}
+
+#[derive(Debug, Deserialize)]
+#[serde(rename_all = "camelCase", deny_unknown_fields)]
+pub(super) struct DeleteOne {
+    filter: Document,
+    session: Option<String>,
+    #[serde(flatten)]
+    options: DeleteOptions,
+}
+
+impl TestOperation for DeleteOne {
+    fn execute_entity_operation<'a>(
+        &'a self,
+        id: &'a str,
+        test_runner: &'a TestRunner,
+    ) -> BoxFuture<'a, Result<Option<Entity>>> {
+        async move {
+            let collection = test_runner.get_collection(id).await;
+            let result = with_opt_session!(
+                test_runner,
+                &self.session,
+                collection
+                    .delete_one(self.filter.clone())
+                    .with_options(self.options.clone()),
+            )
+            .await?;
+            let result = to_bson(&result)?;
+            Ok(Some(result.into()))
+        }
+        .boxed()
+    }
+}
+
